patch

+++ app/src/main/java/apc/appcradle/kotlinjc_friendsactivity_app/domain/model/AppState.kt
@@ -16,7 +16,7 @@ data class AppState(
     //Settings
     val userStepLength: Double = 0.4,
     val currentTheme: AppThemes = AppThemes.System,
-    val userScale: Int = 1,
+    val userScale: Float = 1.0f
 
     //Steps
 //    val weeklySteps: Long = 0
